// apps/web/src/api/generated/models/tmdbCastCredit.ts
/**
 * Generated by orval v8.12.3 🍺
 * Do not edit manually.
 * Verdict API
 * OpenAPI documentation for the Verdict backend.
 * OpenAPI spec version: 1.0.0
 */

export interface TmdbCastCredit {
  adult: boolean;
  gender: number;
  id: number;
  known_for_department: string;
  name: string;
  original_name: string;
  popularity: number;
  /** @nullable */
  profile_path: string | null;
  cast_id: number;
  character: string;
  credit_id: string;
  order: number;
}