// apps/web/src/api/generated/models/tmdbCrewCredit.ts
/**
 * Generated by orval v8.12.3 🍺
 * Do not edit manually.
 * Verdict API
 * OpenAPI documentation for the Verdict backend.
 * OpenAPI spec version: 1.0.0
 */

export interface TmdbCrewCredit {
  adult: boolean;
  gender: number;
  id: number;
  known_for_department: string;
  name: string;
  original_name: string;
  popularity: number;
  /** @nullable */
  profile_path: string | null;
  credit_id: string;
  department: string;
  job: string;
}