// apps/web/src/api/generated/models/tmdbMovieCredits.ts
/**
 * Generated by orval v8.12.3 🍺
 * Do not edit manually.
 * Verdict API
 * OpenAPI documentation for the Verdict backend.
 * OpenAPI spec version: 1.0.0
 */
import type { TmdbCastCredit } from './tmdbCastCredit';
import type { TmdbCrewCredit } from './tmdbCrewCredit';

export interface TmdbMovieCredits {
  id: number;
  cast: TmdbCastCredit[];
  crew: TmdbCrewCredit[];
}