// apps/web/src/api/generated/models/searchMoviesLocale.ts
/**
 * Generated by orval v8.12.3 🍺
 * Do not edit manually.
 * Verdict API
 * OpenAPI documentation for the Verdict backend.
 * OpenAPI spec version: 1.0.0
 */

export type SearchMoviesLocale =
  (typeof SearchMoviesLocale)[keyof typeof SearchMoviesLocale];

export const SearchMoviesLocale = {
  en: 'en',
  ru: 'ru',
} as const;